+++ src/main/java/net/frozenblock/lib/renderer/FrozenLibRenderPipelines.java
@@ -26,6 +26,7 @@ import java.util.Optional;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.frozenblock.lib.FrozenLibConstants;
+import net.minecraft.client.renderer.BindGroupLayouts;
 import net.minecraft.client.renderer.RenderPipelines;
 
 @Environment(EnvType.CLIENT)
@@ -35,7 +36,8 @@ public final class FrozenLibRenderPipelines {
 			.withLocation(FrozenLibConstants.id("pipeline/entity_cutout_no_shading"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
 			.withShaderDefine("NO_CARDINAL_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
+			.withCull(false)
 			.build()
 	);
 
@@ -44,8 +46,8 @@ public final class FrozenLibRenderPipelines {
 			.withLocation(FrozenLibConstants.id("pipeline/entity_cutout_no_shading_cull"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
 			.withShaderDefine("NO_CARDINAL_LIGHTING")
-			.withSampler("Sampler1")
-			.withCull(true)
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
+			.withCull(false)
 			.build()
 	);
 
@@ -54,19 +56,19 @@ public final class FrozenLibRenderPipelines {
 			.withLocation(FrozenLibConstants.id("pipeline/entity_translucent_no_shading_cull"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
 			.withShaderDefine("NO_CARDINAL_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
 			.withColorTargetState(new ColorTargetState(BlendFunction.TRANSLUCENT))
 			.withCull(true)
 			.build()
 	);
 
+	// TODO: check if this functions as intended (depth stencil)
 	public static final RenderPipeline ENTITY_TRANSLUCENT_EMISSIVE_FIXED = RenderPipelines.register(
-		RenderPipeline.builder(RenderPipelines.ENTITY_SNIPPET)
+		RenderPipeline.builder(RenderPipelines.ENTITY_EMISSIVE_SNIPPET)
 			.withLocation(FrozenLibConstants.id("pipeline/entity_translucent_emissive_fixed"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
-			.withShaderDefine("EMISSIVE")
 			.withShaderDefine("PER_FACE_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
 			.withColorTargetState(new ColorTargetState(BlendFunction.TRANSLUCENT))
 			.withCull(false)
 			.withDepthStencilState(DepthStencilState.DEFAULT)
@@ -74,25 +76,23 @@ public final class FrozenLibRenderPipelines {
 	);
 
 	public static final RenderPipeline ENTITY_TRANSLUCENT_EMISSIVE_CULL = RenderPipelines.register(
-		RenderPipeline.builder(RenderPipelines.ENTITY_SNIPPET)
+		RenderPipeline.builder(RenderPipelines.ENTITY_EMISSIVE_SNIPPET)
 			.withLocation(FrozenLibConstants.id("pipeline/entity_translucent_emissive_cull"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
-			.withShaderDefine("EMISSIVE")
 			.withShaderDefine("PER_FACE_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
 			.withColorTargetState(new ColorTargetState(BlendFunction.TRANSLUCENT))
 			.withCull(true)
-			.withDepthStencilState(Optional.empty())
+			.withDepthStencilState(new DepthStencilState(CompareOp.GREATER_THAN_OR_EQUAL, false))
 			.build()
 	);
 
 	public static final RenderPipeline ENTITY_TRANSLUCENT_EMISSIVE_FIXED_CULL = RenderPipelines.register(
-		RenderPipeline.builder(RenderPipelines.ENTITY_SNIPPET)
+		RenderPipeline.builder(RenderPipelines.ENTITY_EMISSIVE_SNIPPET)
 			.withLocation(FrozenLibConstants.id("pipeline/entity_translucent_emissive_fixed_cull"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
-			.withShaderDefine("EMISSIVE")
 			.withShaderDefine("PER_FACE_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
 			.withColorTargetState(new ColorTargetState(BlendFunction.TRANSLUCENT))
 			.withCull(true)
 			.withDepthStencilState(DepthStencilState.DEFAULT)
@@ -100,38 +100,41 @@ public final class FrozenLibRenderPipelines {
 	);
 
 	public static final RenderPipeline ENTITY_TRANSLUCENT_EMISSIVE_ALWAYS_RENDER = RenderPipelines.register(
-		RenderPipeline.builder(RenderPipelines.ENTITY_SNIPPET)
+		RenderPipeline.builder(RenderPipelines.ENTITY_EMISSIVE_SNIPPET)
 			.withLocation(FrozenLibConstants.id("pipeline/entity_translucent_emissive_always_render"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
 			.withShaderDefine("EMISSIVE")
 			.withShaderDefine("PER_FACE_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
 			.withColorTargetState(new ColorTargetState(BlendFunction.TRANSLUCENT))
+			// TODO: test
 			.withDepthStencilState(Optional.empty())
 			.withCull(false)
 			.build()
 	);
 
 	public static final RenderPipeline ENTITY_TRANSLUCENT_EMISSIVE_ALWAYS_RENDER_CULL = RenderPipelines.register(
-		RenderPipeline.builder(RenderPipelines.ENTITY_SNIPPET)
+		RenderPipeline.builder(RenderPipelines.ENTITY_EMISSIVE_SNIPPET)
 			.withLocation(FrozenLibConstants.id("pipeline/entity_translucent_emissive_always_render_cull"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
 			.withShaderDefine("EMISSIVE")
 			.withShaderDefine("PER_FACE_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
 			.withColorTargetState(new ColorTargetState(BlendFunction.TRANSLUCENT))
+			// TODO: test
 			.withDepthStencilState(Optional.empty())
 			.withCull(true)
 			.build()
 	);
 
+	// TODO: test
 	public static final RenderPipeline APPARITION_OUTER = RenderPipelines.register(
 		RenderPipeline.builder(RenderPipelines.ENTITY_SNIPPET)
 			.withLocation(FrozenLibConstants.id("pipeline/apparition_outer"))
 			.withShaderDefine("ALPHA_CUTOUT", 0.1F)
 			.withShaderDefine("EMISSIVE")
 			.withShaderDefine("NO_CARDINAL_LIGHTING")
-			.withSampler("Sampler1")
+			.withBindGroupLayout(BindGroupLayouts.SAMPLER1)
 			.withColorTargetState(new ColorTargetState(BlendFunction.TRANSLUCENT))
 			.withCull(true)
 			.withDepthStencilState(new DepthStencilState(CompareOp.LESS_THAN_OR_EQUAL, false))